+++ provd/config.py
@@ -23,13 +23,16 @@ The following parameters are defined:
         info_extractor
         retriever
         updater
-        external_ip
-        http_port
         tftp_port
         verbose
         sync_service_type
         asterisk_ami_servers
-        num_http_proxies
+        advertised_host
+            The hostname of the provisioning server advertised to phones
+        advertised_http_port
+            The HTTP port advertised to phones
+        advertised_http_url
+            The HTTP URL advertised to phones
     rest_api:
         ip
         port
@@ -70,7 +73,6 @@ from __future__ import annotations
 import logging
 import json
 import os.path
-import socket
 from typing import Any, TypedDict, Union, cast, Literal
 
 from twisted.python import usage
@@ -93,9 +95,12 @@ class SyncDbConfigDict(TypedDict):
 
 
 class GeneralConfig(TypedDict):
-    external_ip: str
-    listen_interface: str
-    listen_port: int
+    http_proxied_listen_interface: str
+    http_proxied_listen_port: int
+    http_proxied_trusted_proxies_count: int
+    advertised_host: Union[str, None]
+    advertised_http_url: Union[str, None]
+    advertised_http_port: int
     base_raw_config: dict[str, Any]
     base_raw_config_file: str
     request_config_dir: str
@@ -108,14 +113,11 @@ class GeneralConfig(TypedDict):
     info_extractor: str
     retriever: str
     updater: str
-    http_port: int
     tftp_port: int
-    base_external_url: str | None
     verbose: bool
     sync_service_type: str
-    num_http_proxies: int
     syncdb: SyncDbConfigDict
-    http_auth_strategy: Literal['url_key']
+    http_auth_strategy: Union[Literal['url_key'], None]
 
 
 class RestApiConfigDict(TypedDict):
@@ -167,20 +169,18 @@ class ProvdConfigDict(TypedDict):
     amid: AmidConfigDict
     bus: BusConfigDict
     plugin_config: dict[str, Any]
+    tenants: dict[str, dict]
 
 
 logger = logging.getLogger(__name__)
-
-DEFAULT_LISTEN_PORT = 8667
-DEFAULT_HTTP_PORT = 8667
 
 _DEFAULT_CONFIG: ProvdConfigDict = {
     'config_file': '/etc/wazo-provd/config.yml',
     'extra_config_files': '/etc/wazo-provd/conf.d',
     'general': {
-        'external_ip': '127.0.0.1',
-        'listen_interface': '0.0.0.0',
-        'listen_port': DEFAULT_LISTEN_PORT,
+        'advertised_host': '127.0.0.1',
+        'advertised_http_port': 8667,
+        'advertised_http_url': None,
         'base_raw_config': {},
         'base_raw_config_file': '/etc/wazo-provd/base_raw_config.json',
         'request_config_dir': '/etc/wazo-provd',
@@ -193,15 +193,12 @@ _DEFAULT_CONFIG: ProvdConfigDict = {
         'info_extractor': 'default',
         'retriever': 'default',
         'updater': 'default',
-        'http_port': DEFAULT_HTTP_PORT,
         'tftp_port': 69,
         'http_proxied_listen_interface': '127.0.0.1',
         'http_proxied_listen_port': 18667,
         'http_proxied_trusted_proxies_count': 1,
-        'base_external_url': None,
         'verbose': False,
         'sync_service_type': 'none',
-        'num_http_proxies': 0,
         'syncdb': {
             'interval_sec': 86400,
             'start_sec': 60,
@@ -243,13 +240,14 @@ _DEFAULT_CONFIG: ProvdConfigDict = {
         'exchange_type': 'headers',
     },
     'plugin_config': {},
+    'tenants': {},
 }
 
 _OPTION_TO_PARAM_LIST = [
     # (<option name, (<section, param name>)>)
     ('config-file', ('general', 'config_file')),
     ('config-dir', ('general', 'request_config_dir')),
-    ('http-port', ('general', 'http_port')),
+    ('http-port', ('general', 'advertised_http_port')),
     ('tftp-port', ('general', 'tftp_port')),
     ('rest-port', ('general', 'rest_port')),
 ]
@@ -301,12 +299,6 @@ def _load_json_file(raw_value: str) -> dict[str, Any]:
         return json.load(f)
 
 
-def _process_aliases(raw_config: ProvdConfigDict) -> None:
-    if 'ip' in raw_config['general'] and 'external_ip' not in raw_config['general']:
-        ip = raw_config['general']['ip']  # type: ignore[typeddict-item]
-        raw_config['general']['external_ip'] = ip
-
-
 def _check_and_convert_parameters(raw_config: dict[str, Any]) -> None:
     if raw_config['rest_api']['ssl']:
         if 'ssl_certfile' not in raw_config['rest_api']:
@@ -322,33 +314,22 @@ def _check_and_convert_parameters(raw_config: dict[str, Any]) -> None:
     )
 
 
-def _get_ip_fallback():
-    # This function might return an IP address of a loopback interface, but we
-    # don't care since it's not possible to determine implicitly which IP address
-    # we should use anyway.
-    return socket.gethostbyname(socket.gethostname())
-
-
 def _update_general_base_raw_config(app_raw_config: dict[str, Any]) -> None:
     # warning: raw_config in the function name means device raw config and
     # the app_raw_config argument means application configuration.
     base_raw_config = app_raw_config['general']['base_raw_config']
     base_raw_config |= {
-        'http_port': app_raw_config['general']['http_port'],
+        'http_port': app_raw_config['general']['advertised_http_port'],
         'tftp_port': app_raw_config['general']['tftp_port'],
     }
-    if app_raw_config['general'].get('base_external_url'):
+    if app_raw_config['general']['advertised_http_url']:
         base_raw_config |= {
-            'http_base_url': app_raw_config['general']['base_external_url'],
+            'http_base_url': app_raw_config['general']['advertised_http_url'],
         }
 
     if 'ip' not in base_raw_config:
-        if 'external_ip' in app_raw_config['general']:
-            external_ip = app_raw_config['general']['external_ip']
-        else:
-            external_ip = _get_ip_fallback()
-            logger.warning('Using "%s" for base raw config ip parameter', external_ip)
-        base_raw_config['ip'] = external_ip
+        advertised_host = app_raw_config['general']['advertised_host']
+        base_raw_config['ip'] = advertised_host
 
 
 def _post_update_raw_config(raw_config: dict[str, Any]) -> None:
@@ -380,7 +361,6 @@ def get_config(argv: Options) -> ProvdConfigDict:
     file_config = read_config_file_hierarchy(ChainMap(cli_config, _DEFAULT_CONFIG))
     service_key = _load_key_file(ChainMap(cli_config, file_config, _DEFAULT_CONFIG))
     raw_config = ChainMap(cli_config, service_key, file_config, _DEFAULT_CONFIG)
-    _process_aliases(raw_config)
     _check_and_convert_parameters(raw_config)
     _post_update_raw_config(raw_config)
     return cast(ProvdConfigDict, raw_config)