+++ provd/devices/ident.py
@@ -50,7 +50,7 @@ class ExtractorProtocol(Protocol):
         ...
 
 
-class DevTFTPRequest(TFTPRequest, total=False):
+class DevTFTPRequest(TFTPRequest):
     prov_dev: DeviceInfo
 
 
@@ -863,14 +863,18 @@ class HTTPKeyVerifyingHook(BaseHTTPHookService):
         self._app = app
         super().__init__(*args, **kwargs)
 
-    def getChild(self, path: str, request: Request):
-        logger.debug('URL key verifying hook, path = "%s", request = "%s"', path, request)
+    def getChild(self, path: bytes, request: Request):
+        logger.debug(
+            'URL key verifying hook, path = "%s", request = "%s"', path, request
+        )
         prov_key = path.decode('utf-8')
         logger.debug('Prov key = "%s"', prov_key)
         if not prov_key:
             logger.info('Empty provisioning key. Denying request.')
             return self.unauthorized_resource
-        tenant_uuid = self._app.configure_service.get_tenant_from_provisioning_key(prov_key)
+        tenant_uuid = self._app.configure_service.get_tenant_from_provisioning_key(
+            prov_key
+        )
         if not tenant_uuid:
             logger.info('Invalid URL key. Denying request.')
             return self.unauthorized_resource
@@ -885,9 +889,13 @@ class HTTPKeyVerifyingHook(BaseHTTPHookService):
 
         path = request.postpath.pop(0)
         request.prepath.append(path)
-        logger.debug('Rewritten request: path = %s, prepath = %s, postpath = %s', path, request.prepath, request.postpath)
+        logger.debug(
+            'Rewritten request: path = %s, prepath = %s, postpath = %s',
+            path,
+            request.prepath,
+            request.postpath,
+        )
         return self._next_service(path, request)
-
 
 
 # @implementer(ITFTPReadService)